requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyJWT
django-cors-headers